Move this handy matcher somewhere reusable
